# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: ROBBED_ API
  version: 1.0.0
  description: |
    Frozen HTTP contract for `apps/api` (Hono on Bun) — transcribed from
    `docs/services/api.md` (Design v1.0) and `launchpad-spec.md` §2/§2.1/§5/§8/§12.
    M2 implements this document; M3 (frontend) consumes it. Response DTO shapes
    are the zod schemas in `packages/shared/src/api-types.ts` — the schemas
    below mirror them and MUST NOT drift (shared package is authoritative).

    Conventions (api.md §2):
    - Envelope: `{ data, error: null }` on success, `{ data: null, error: { code, message } }` on failure.
    - All uint256 values are decimal strings.
    - Every event-derived object carries `confirmationState` (spec §2.1): `soft_confirmed | posted_to_l1 | finalized`.
    - Every USD-derived field is `{ usd, ethUsd, asOf }` computed at request time from `eth_usd_snapshots`
      (never a constant, spec §2); `stale: true` added when the snapshot is older than 5 minutes.
    - List endpoints use cursor pagination (`?cursor=&limit=`, limit ≤ 100, default 50).
    - Listing-gated endpoints (`/tokens` list, `/search`) exclude `visibility = 'hidden'`;
      direct `/tokens/{address}` returns hidden tokens with the moderation flag, never 404 (spec §12.21).
    - Rate limiting (api.md §6.3): 429 with `Retry-After`. Defaults: uploads 10/h + 3/min per IP,
      metadata 20/h per IP, search 60/min per IP, reads 300/min per IP, admin 60/min per session.
...
  license:
    name: MIT
servers:
  - url: /
...
security: []
...
paths:
...
  /v1/metadata:
    post:
      tags: [launch]
      operationId: createMetadata
      summary: Canonicalize + hash + publish metadata JSON (spec §8.3, §12.19)
      description: |
        Server builds the fixed field set + version tag, canonicalizes with the shared
        `canonicalizeMetadata` (packages/shared — the single implementation), keccak256-hashes,
        and PUTs to R2 at `metadata/{hash}.json` (content-addressed). Text moderation +
        impersonation checks run but NEVER block the response (moderation gates listing,
        not creation — spec §8.4). The client MUST re-verify `metadataHash` with the shared
        canonicalizer before signing the tx (normative for M3 — spec §12.19).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MetadataRequest" }
      responses:
        "200":
          description: Canonical metadata published
          content:
            application/json:
              schema:
                type: object
                required: [data, error]
                properties:
                  data: { $ref: "#/components/schemas/MetadataResponse" }
                  error: { type: "null" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/search:
    get:
      tags: [search]
      operationId: search
      summary: Search over name, ticker, contract address, creator address (spec §5.1)
      description: |
        Single endpoint, pg_trgm (api.md §3.3). `^0x[0-9a-fA-F]{6,40}$` queries switch to
        address mode (prefix/exact on token + creator address, exact pins first); otherwise
        trigram similarity over name/ticker with ticker boost ×1.2, volume tiebreak,
        similarity floor 0.25 (ratified defaults, spec §12.22). Hidden listings excluded.
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 80 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Matching tokens (same card projection as /v1/tokens)
          content:
            application/json:
              schema:
                type: object
                required: [data, error]
                properties:
                  data:
                    type: object
                    required: [results]
                    properties:
                      results:
                        type: array
                        items: { $ref: "#/components/schemas/TokenCard" }
                  error: { type: "null" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/tokens:
    get:
      tags: [tokens]
      operationId: listTokens
      summary: Token grid — sorts + filters (spec §5.1)
      description: |
        `trending = vol24h × e^(−age/24h)` (ratified default, spec §12.22, tunable config);
        `progress = real_eth_reserves / graduation_eth` (pre-grad only). Hidden listings excluded;
        `pending_review` remains listed (spec §12.21).
      parameters:
        - name: sort
          in: query
          schema:
            type: string
            enum: [trending, newest, mcap, volume24h, progress]
            default: trending
        - name: filter
          in: query
          schema:
            type: string
            enum: [pregrad, graduated, all]
            default: all
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Token cards
          content:
            application/json:
              schema:
                type: object
                required: [data, error]
                properties:
                  data:
                    type: object
                    required: [tokens, nextCursor]
                    properties:
                      tokens:
                        type: array
                        items: { $ref: "#/components/schemas/TokenCard" }
                      nextCursor: { $ref: "#/components/schemas/NextCursor" }
                  error: { type: "null" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /v1/stats:
    get:
      tags: [tokens]
      operationId: getStats
      summary: Global stats (api.md §3.4)
      responses:
        "200":
          description: Tokens launched, graduations, 24h volume, treasury fees — all computed, USD per §2
          content:
            application/json:
              schema:
                type: object
                required: [data, error]
                properties:
                  data: { $ref: "#/components/schemas/StatsResponse" }
                  error: { type: "null" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/confirmations:
    get:
      tags: [meta]
      operationId: getConfirmations
      summary: Confirmation watermarks (spec §2.1; SSR initial state)
      responses:
        "200":
          description: Current safe/finalized/latest L2 blocks
          content:
            application/json:
              schema:
                type: object
                required: [data, error]
                properties:
                  data: { $ref: "#/components/schemas/ConfirmationsResponse" }
                  error: { type: "null" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/eth-usd:
    get:
      tags: [meta]
      operationId: getEthUsd
      summary: ETH/USD snapshot with source + asOf (spec §2 — never a constant)
      responses:
        "200":
          description: Latest snapshot from eth_usd_snapshots (indexer poller writes, API reads)
          content:
            application/json:
              schema:
                type: object
                required: [data, error]
                properties:
                  data: { $ref: "#/components/schemas/EthUsdResponse" }
                  error: { type: "null" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: robbed_admin_session
      description: |
        SIWE (EIP-4361) against a config address allowlist → short-lived signed session
        cookie (HttpOnly, SameSite=strict, 12h). Mutations additionally require the
        `X-CSRF-Token` header (api.md §6.2). Cookie name is implementation-defined.
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: Opaque cursor from a previous response (api.md §2)
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
...
  responses:
    BadRequest:
      description: Validation failure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    RateLimited:
      description: Rate limit exceeded (api.md §6.3)
      headers:
        Retry-After:
          description: Seconds until the sliding window admits the client again
          schema: { type: integer, minimum: 0 }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
...
    Address:
      type: string
      pattern: "^0x[0-9a-fA-F]{40}$"
      description: EVM address (responses use lowercase)
    Hex32:
      type: string
      pattern: "^0x[0-9a-f]{64}$"
      description: 0x-prefixed 32-byte hex (tx hash / keccak256)
    DecimalString:
      type: string
      pattern: "^[0-9]+$"
      description: uint256 as decimal string (api.md §2 — exceeds JS safe integer)
    NextCursor:
      oneOf:
        - type: string
        - type: "null"
      description: Opaque cursor; null when exhausted
    ConfirmationState:
      type: string
      enum: [soft_confirmed, posted_to_l1, finalized]
      description: Spec §2.1 — soft-confirmed → posted-to-L1 → finalized (monotonic)
...
    ModerationVisibility:
      type: string
      enum: [visible, pending_review, hidden]
    TokenStatus:
      type: string
      enum: [curve, graduating, graduated]
      description: Derived (indexer.md §3.2) — drives the venue pill / trade-widget engine
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [oversized, unsupported_type, decode_failed, rate_limited, not_found, invalid_request, unauthorized, upstream_unavailable, conflict]
          description: >-
            Closed set — single source is packages/shared errorCodeSchema
            (api-types.ts). A new code is added there and here in lockstep.
            `upstream_unavailable` = unexpected 500 / readyz-503 dependency-down;
            `conflict` = request conflicts with stored state (e.g. unknown imageHash).
        message: { type: string }
    ErrorEnvelope:
      type: object
      required: [data, error]
      properties:
        data: { type: "null" }
        error: { $ref: "#/components/schemas/Error" }

    # ── Value objects ────────────────────────────────────────────────────────
    UsdValue:
      type: object
      required: [usd, ethUsd, asOf]
      properties:
        usd: { type: string }
        ethUsd: { type: string }
        asOf: { type: string, format: date-time }
        stale:
          type: boolean
          const: true
          description: Present iff the ETH/USD snapshot is older than 5 minutes (api.md §2)
      description: Computed at request time from eth_usd_snapshots — never a constant (spec §2)
    ModerationInfo:
      type: object
      required: [visibility, impersonationFlag]
      properties:
        visibility: { $ref: "#/components/schemas/ModerationVisibility" }
        impersonationFlag: { type: boolean }
        impersonationTicker:
          type: string
          description: Present on TokenDetail when impersonation-flagged (api.md §3.4)

    # ── Core DTOs (mirror packages/shared/src/api-types.ts) ─────────────────
    TokenCard:
      type: object
      required:
        [address, name, ticker, imageUrl, creator, createdAt, priceEth, mcap,
         progressPct, change24hPct, volume24h, graduated, status,
         confirmationState, moderation]
      properties:
        address: { $ref: "#/components/schemas/Address" }
        name: { type: string }
        ticker: { type: string }
        imageUrl:
          oneOf: [{ type: string }, { type: "null" }]
          description: Null until metadata JSON fetched (indexer.md §3.1)
        creator: { $ref: "#/components/schemas/Address" }
        createdAt: { type: integer, description: "block timestamp, unix seconds" }
        priceEth:
          oneOf: [{ type: number }, { type: "null" }]
          description: Display-only float; null before first trade
        mcap: { $ref: "#/components/schemas/UsdValue" }
        mcapEth:
          allOf: [{ $ref: "#/components/schemas/DecimalString" }]
          description: >-
            Native ETH market cap, wei decimal string (decisions.md §7.2 item 3).
            ETH-first display source (§2) — USD mcap derives from it. Optional /
            additive: absent until the indexer materializes it; not in `required`.
        progressPct: { type: number, description: real_eth_reserves / graduation_eth }
        change24hPct:
          oneOf: [{ type: number }, { type: "null" }]
        volume24h: { $ref: "#/components/schemas/DecimalString" }
        graduated: { type: boolean }
        status: { $ref: "#/components/schemas/TokenStatus" }
        confirmationState: { $ref: "#/components/schemas/ConfirmationState" }
        moderation: { $ref: "#/components/schemas/ModerationInfo" }
...
    StatsResponse:
      type: object
      required:
        [tokensLaunched, graduations, volume24hEth, volume24h,
         treasuryFeesCollectedWeth, treasuryFeesCollected]
      properties:
        tokensLaunched: { type: integer, minimum: 0 }
        graduations: { type: integer, minimum: 0 }
        volume24hEth: { $ref: "#/components/schemas/DecimalString" }
        volume24h: { $ref: "#/components/schemas/UsdValue" }
        treasuryFeesCollectedWeth: { $ref: "#/components/schemas/DecimalString" }
        treasuryFeesCollected: { $ref: "#/components/schemas/UsdValue" }
    ConfirmationsResponse:
      type: object
      required: [safeBlock, finalizedBlock, latestBlock, updatedAt]
      properties:
        safeBlock: { type: integer, minimum: 0 }
        finalizedBlock: { type: integer, minimum: 0 }
        latestBlock: { type: integer, minimum: 0 }
        updatedAt: { type: string, format: date-time }
    EthUsdResponse:
      type: object
      required: [price, source, asOf]
      properties:
        price: { type: number }
        source: { type: string, description: "e.g. 'chainlink:4663', 'defillama' — config-driven (spec §13)" }
        asOf: { type: string, format: date-time }
...
    MetadataRequest:
      type: object
      required: [name, ticker, imageUrl, imageHash]
      description: >-
        name/ticker limits are UTF-8 BYTE limits (spec §12.30) matching the
        on-chain gate bytes(x).length — name ≤ 32 bytes, ticker ≤ 10 bytes —
        server-enforced by packages/shared byteBoundedString. JSON-Schema
        maxLength counts Unicode code points, so it CANNOT express a byte bound
        exactly; the maxLength values below are safe upper bounds (documentary),
        and the authoritative gate is the server-side byte-length refinement.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 32
          description: ≤ 32 UTF-8 BYTES (authoritative), server-enforced (§12.30)
        ticker:
          type: string
          minLength: 1
          maxLength: 10
          description: ≤ 10 UTF-8 BYTES (authoritative), server-enforced (§12.30)
        description: { type: string, maxLength: 500 }
        links:
          type: object
          properties:
            website: { type: string, format: uri }
            x: { type: string, format: uri }
            telegram: { type: string, format: uri }
        imageUrl: { type: string, format: uri, description: must be our CDN origin }
        imageHash:
          allOf: [{ $ref: "#/components/schemas/Hex32" }]
          description: must match an object the upload endpoint produced (R2 existence check)
    MetadataResponse:
      type: object
      required: [metadataHash, metadataUri, canonicalJson]
      properties:
        metadataHash:
          allOf: [{ $ref: "#/components/schemas/Hex32" }]
          description: keccak256 of the canonical bytes — the client MUST re-verify with the shared canonicalizer before signing (spec §12.19)
        metadataUri: { type: string, description: "R2 URL metadata/{hash}.json (content-addressed)" }
        canonicalJson: { type: string, description: the exact canonical JSON text that was hashed and stored }

    # ── Admin ────────────────────────────────────────────────────────────────
...